requirements.txt
django<=1.11
djangorestframework
pillow
argon2-cffi
